Added mandatory `data_length` parameter to `user_defined()` constructor.
Since there is no documentation, it is very easy to miss that the
msg_length header is supposed to be set to the length of header +
message. By default, it is set to the header length only.

Forgetting to set the message length results in just messages being
truncated and causing errors that are hard to interpret. Adding the
`data_length` parameter at least forces that field to be set.